Separated compilation and globals to individual files. Fixed instruction overflow.
