+++ instr_small
@@ -1,0 +1,25 @@
+LDR X0 2
+LDR X1 5
+ADD X2 X0 X1
+PSH 1
+STR X2
+PSH 3
+PSH 5
+PSH 10
+PSH 100
+PSH 23
+PSH 834
+PSH 12389
+PSH 280
+LDR X3 1
+SUB X7 X2 X3
+STR X7
+SAV
+BR func
+END
+-func
+LDR X5 10
+LDR X6 10
+ADD X5 X5 X6
+EX
+*
